pyproject.toml
[project]
name = "retriever"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "pydantic>=2.11.5",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
]
